Revert "feat(BANK-8): Criação de pastas para APPs, DTOs, Entitys
