+++ backend/src/main.ts
@@ -1,5 +1,5 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { AppModule } from './APPs/app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
